Fix : 평균값 전달 실수 형식으로 변경

+++ backend/src/main/java/com/ssafy/backend/domain/analysis/dto/response/GithubVsInfo.java
@@ -15,8 +15,8 @@ import lombok.Getter;
 public class GithubVsInfo {
 	String nickname;
 	String avatarUrl;
-	int commit;
-	int star;
+	double commit;
+	double star;
 	long repositories;
 	List<GithubDetailLanguage> languages;
